pyproject.toml
[tool.poetry]
name = "django-uml"
version = "0.1.0"
description = ""
authors = [
    "Thijs Baaijen <Thijss@users.noreply.github.com>",
]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
django-extensions = "^3.1.0"
pygraphviz = "^1.6"
django-environ = "^0.4.5"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"
pylint-django = "^2.4.2"
pycodestyle = "^2.6.0"
coverage = "^5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'MASTER']
load-plugins = "pylint_django"
django-settings-module="config.settings.default"

[tool.coverage.report]
omit = [
  "*/config/settings*",
]
